[Bugfix] Front-end and back-end overhaul
First, DO NOT BE ALARMED, no part of meeting proposals were removed. The link in the navbar however was. Why? Because the feature is broken at the moment anyway. It looks like it needs to be redesigned from scratch to operate correctly with a separate proposed-time-model. When that is done it can be attached to on-site and email notifications.

Backend overhaul:
* All controllers have had their logic reviewed, and appropriate adjustments have been made.
* Max file size available for requests (profile images and post attachments) has been increased from 4MB to 100MB.
* IdentityUser has been extended with a DisplayName claim, and it is now used to display the users name on the navbar link to the profile. This is due to the fact that a display name is a lot shorter than their email.

Frontend overhaul: This patch introduces the FontAwesome API in _Layout.cshtml, and icons have been implemented all over.